Add files via upload
EXERCISES: 

Create a python script where the user is going to be asked for the name. After that, the program will greet the user with the given name. Only exception: the script will not greet the user if the given name is "Ana", "John" or "Peter".

Create a function in a pyhton script that returns (print) the last element of a given list, reverts the order of the list and finally returns (print) the reverted list. 

Create a function in a python script that takes a list of strings as input and then it prints, line by line, the whole list in a rectangular shape. E.g., the list could be ["Hi", "my", "name", "is", "John", "Doe"] and the final result is:

********** 
* Hi        * 
* my       *
* name   * 
* is         * 
* John    * 
* Doe     *
**********